Break Window utility functions into Window.c,h.
