+++ Sources/ObjectivelyMVC.h
@@ -37,12 +37,12 @@
 #include <ObjectivelyMVC/Colors.h>
 #include <ObjectivelyMVC/Constraint.h>
 #include <ObjectivelyMVC/Control.h>
-#include <ObjectivelyMVC/Input.h>
 #include <ObjectivelyMVC/Font.h>
 #include <ObjectivelyMVC/HSVColorPicker.h>
 #include <ObjectivelyMVC/HueColorPicker.h>
 #include <ObjectivelyMVC/Image.h>
 #include <ObjectivelyMVC/ImageView.h>
+#include <ObjectivelyMVC/Input.h>
 #include <ObjectivelyMVC/Label.h>
 #include <ObjectivelyMVC/Log.h>
 #include <ObjectivelyMVC/NavigationViewController.h>
@@ -71,4 +71,5 @@
 #include <ObjectivelyMVC/View.h>
 #include <ObjectivelyMVC/View+JSON.h>
 #include <ObjectivelyMVC/ViewController.h>
+#include <ObjectivelyMVC/Window.h>
 #include <ObjectivelyMVC/WindowController.h>
