+++ Sources/ObjectivelyMVC/Colors.h
@@ -23,8 +23,6 @@
 
 #pragma once
 
-#include <SDL2/SDL_pixels.h>
-
 #include <ObjectivelyMVC/Types.h>
 
 /**
